Feature/#26 ci upgrade (#27)
* build and test working with Gradle 6 and Java 14 locally

* upgrade to Java 11, Gradle 6, and JitPack to get Opt4J

* added codacy ignores, removed travis and old mercurial

* Update gradle.yml

+++ opendse-io/src/test/java/net/sf/opendse/io/SpecificationWriterAnonymizedTest.java
@@ -65,10 +65,10 @@ public class SpecificationWriterAnonymizedTest {
 		Assert.assertNotNull(anonymized.getEdge("dependency0"));
 
 		Task anonymizedTask = anonymized.getVertex("task0");
-		Assert.assertEquals(anonymizedTask.getAttribute("cost"), 25);
+		Assert.assertEquals((int) anonymizedTask.getAttribute("cost"), 25);
 
 		Dependency anonymizedDependency = anonymized.getEdge("dependency0");
-		Assert.assertEquals(anonymizedDependency.getAttribute("test"), 0);
+		Assert.assertEquals((int) anonymizedDependency.getAttribute("test"), 0);
 
 		Assert.assertNull(anonymized.getVertex("t1"));
 		Assert.assertNull(anonymized.getVertex("t2"));
@@ -101,10 +101,10 @@ public class SpecificationWriterAnonymizedTest {
 		Assert.assertNotNull(anonymized.getEdge("link0"));
 
 		Resource anonymizedResource = anonymized.getVertex("resource0");
-		Assert.assertEquals(anonymizedResource.getAttribute("cost"), 25);
+		Assert.assertEquals((int) anonymizedResource.getAttribute("cost"), 25);
 
 		Link anonymizedLink = anonymized.getEdge("link0");
-		Assert.assertEquals(anonymizedLink.getAttribute("test"), 0);
+		Assert.assertEquals((int) anonymizedLink.getAttribute("test"), 0);
 
 		Assert.assertNull(anonymized.getVertex("r1"));
 		Assert.assertNull(anonymized.getVertex("r2"));
